modify dealer update method

+++ app/Http/Controllers/UserController.php
@@ -3,12 +3,15 @@
 namespace App\Http\Controllers;
 
 use App\User;
-use Illuminate\Http\Request;
 
 class UserController extends Controller
 {
     public function destroy(User $user)
     {
+        if ($user->employee->permission_id == 1) {
+            return redirect('/permission-denied');
+        }
+
         $user->forceDelete();
 
         return redirect()->back()->with('deleted', 'Deleted Successfully');
